feat(billing): remove int64

+++ billing.go
@@ -82,7 +82,7 @@ type Billing struct {
 	FirstSix          string `xml:"first_six,omitempty"`
 	LastFour          string `xml:"last_four,omitempty"`
 	CardType          string `xml:"card_type,omitempty"`
-	Number            int64  `xml:"number,omitempty"`
+	Number            int    `xml:"number,omitempty"`
 	Month             int    `xml:"month,omitempty"`
 	Year              int    `xml:"year,omitempty"`
 	VerificationValue int    `xml:"verification_value,omitempty"` // Create/update only
